create more universal `SortableInfiniteList`
